backend/requirements.txt
bcrypt==4.1.3
blinker==1.8.2
certifi==2024.6.2
click==8.1.7
cloudinary==1.40.0
colorama==0.4.6
Flask==3.0.3
Flask-Login==0.6.3
flask-marshmallow==1.2.1
Flask-SQLAlchemy==3.1.1
flask-cors==4.0.0
greenlet==3.0.3
gunicorn==22.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
marshmallow==3.21.3
marshmallow-sqlalchemy==1.0.0
packaging==24.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
six==1.16.0
SQLAlchemy==2.0.30
typing_extensions==4.12.2
urllib3==2.2.1
Werkzeug==3.0.3
requests>=2.31.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pymongo==4.6.0
openai>=1.35.0
numpy>=1.24.0
python-multipart==0.0.6
python-json-logger==2.0.7